ci: 🎡 fix build errors due to lack of ansi colour support on ci
When secrets are added to Travis CI, ANSI colour support is disabled
because the output has to be filtered. This causes some Jest snapshot
tests, such as `.toThrowErrorMatchingSnapshot()` checks, to become
inconsistent with the snapshots generated on a console that supports
ANSI colour output. This is fixed by running snapshots through a
serialiser that strips ANSI escape sequences.